Ensure resque adapter is being used by Active Job on resque tests
Before this worked without issues because resque was the only adapter set,
but after introducing Solid Queue, depending on the order of the tests, it
might happen that we run a resque test after a solid queue test, that will
finish with the solid queue adapter being used. This is a problem if we try
to fetch jobs of a certain kind via Active Job (eg. ApplicationJob.jobs.pending
and things like that), as that would use Solid Queue.