Don't use a default queue if none is specified
This messes up with failed filtering, and prevents it from working in Solid
Queue. In Solid Queue, failed is not a queue but a status of jobs, that continue
to belong to other queues. In Resque, failed is a separate queue, so when
we use the default queue when querying failed jobs, that's ignored and it
doesn't matter. However, for Solid Queue, we end up showing failed jobs
in the default queue alone, which is not what we want, since right now we
don't have a way to filter failed jobs by queue. We want to show all jobs
in all queues with failed status.

In all the other cases where we filter jobs, we want to provide a queue explicitly,
not the default one, and this is what happens everywhere. Not having a default
saves us from having to explicitly unset the queue every time we want to
fetch failed jobs.