Remove `pending` as default status for job relation
It's much clearer if we force to specify a status to fetch jobs rather
than using `pending` by default.

The only exception for this is when we fetch jobs via a queue: in this case we
assume we want pending jobs. The `queue` object is just a convenience to access
pending jobs in the queue.