Don't reference any actions in JobScoped concern, as it might not exist
Thanks to @excid3 for spotting this. In Rails 7.1, because of the new default
for `config.action_controller.raise_on_missing_callback_actions`, the except
option used in `before_action` for `index` raises an error for those controllers
that don't have an index action. Let's just remove that and let the only controller
with an index action using this concern opt-out of it explicitly.