Convert authentication generator to Rails command
Generators require the whole app to be booted, which involves having
access to the DB, loading all controlers, etc. which we don't need for
just generating the credentials, and might not work when trying to add
production credentials because many apps aren't prepared to boot for
production in another environment.